agent/requirements.txt
botocore
bedrock-agentcore
strands-agents
strands-agents[otel]
strands-agents-tools
aws-opentelemetry-distro
mcp
